requirements.txt
requests==2.31.0
PyGithub==1.59.0
python-gitlab==3.15.0
python-dotenv==1.0.0
tqdm==4.66.1 